fix buy bug

+++ src/main/java/com/netease/course/aop/CacheInterceptor.java
@@ -3,6 +3,8 @@ package com.netease.course.aop;
 import java.util.ArrayList;
 import java.util.List;
 
+import javax.servlet.http.HttpServletRequest;
+
 import org.apache.log4j.Logger;
 import org.aspectj.lang.JoinPoint;
 import org.aspectj.lang.ProceedingJoinPoint;
@@ -18,8 +20,10 @@ public class CacheInterceptor {
 	@Autowired
 	private MemCachedClient memCachedClient;
 
+	
 	public static final int TIMEOUT = 3600;// 缓存失效时间1小时
-	private static final Logger logger=Logger.getLogger(CacheInterceptor.class);
+	private static final Logger logger = Logger.getLogger(CacheInterceptor.class);
+
 	public Object doAround(ProceedingJoinPoint pjp) throws Throwable {
 		// 获取key
 		String cacheKey = getCacheKey(pjp);
@@ -46,27 +50,32 @@ public class CacheInterceptor {
 
 	/**
 	 * 当这个类下某个方法改动数据库时，删除该类的所有缓存
+	 * 
 	 * @param jp
 	 */
 	public void doAfter(JoinPoint jp) {
 		String packageName = jp.getTarget().getClass().getName();
-		
+		System.out.println(packageName);
 		if (!memCachedClient.stats().isEmpty()) {
 			@SuppressWarnings("unchecked")
-			List<String> list =(List<String>) memCachedClient.get(packageName);
-			//遍历
-			for(String key:list){
+			List<String> list = (List<String>) memCachedClient.get(packageName);
+			// 遍历
+			if (list != null) {
+				for (String key : list) {
 					memCachedClient.delete(key);
-
+					System.out.println("删除"+key);
+				}
 			}
-		}else{
+		} else {
 			logger.error("memcached连接失败");
 		}
 
-		//memCachedClient.flushAll();
+		// memCachedClient.flushAll();
 
 	}
-
+	public void doAddList(){
+		memCachedClient.flushAll();
+	}
 	// Key生成规则，包名+ 类名 + 方法名 + 参数
 	public String getCacheKey(ProceedingJoinPoint pjp) {
 
@@ -85,26 +94,29 @@ public class CacheInterceptor {
 
 		return key.toString();
 	}
-/**
- * 将key保存在memcash中
- * 以类名分类，将同类下所有方法的key保存在一起
- * @param cacheKey
- * @param pjp
- */
+
+	/**
+	 * 将key保存在memcash中 以类名分类，将同类下所有方法的key保存在一起
+	 * 
+	 * @param cacheKey
+	 * @param pjp
+	 */
 	public void saveCacheKey(String cacheKey, ProceedingJoinPoint pjp) {
 
 		String packageName = pjp.getTarget().getClass().getName();
-
+		System.out.println(packageName);
 		if (memCachedClient.get(packageName) == null) {
-			List<String> list =new ArrayList<String>();
+			List<String> list = new ArrayList<String>();
 			list.add(cacheKey);
+			System.out.println("新增key:"+cacheKey);
 			memCachedClient.set(packageName, list);
 
 		} else {
 			@SuppressWarnings("unchecked")
 			List<String> list = (List<String>) memCachedClient.get(packageName);
 			list.add(cacheKey);
-			memCachedClient.set(packageName,list);
+			System.out.println("增加key"+cacheKey);
+			memCachedClient.set(packageName, list);
 
 		}
